refactor: LottoNumbers 클래스 제거
