+++ src/main/java/lotto/domain/LottoNumber.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 import java.util.Objects;
 
-public class LottoNumber {
+class LottoNumber {
     private static final int MIN = 1;
     private static final int MAX = 45;
     private final int number;
