+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,11 +1,17 @@
 package lotto.domain;
 
+import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class LottoTest {
     @Test
@@ -21,5 +27,21 @@ class LottoTest {
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
-    // TODO: 추가 기능 구현에 따른 테스트 코드 작성
+    @ParameterizedTest
+    @NullAndEmptySource
+    @MethodSource("testSize")
+    void 중복되지않은_번호가_6개가_아니면_에러를_던진다(List<Integer> numbers) {
+        assertThrows(IllegalArgumentException.class, () -> {
+            new Lotto(numbers);
+        });
+    }
+
+    private static Stream<Arguments> testSize() {
+        return Stream.of(
+                Arguments.of(List.of(1)),
+                Arguments.of(List.of(1, 2, 3, 4, 5)),
+                Arguments.of(List.of(1, 2, 3, 4, 5, 5)),
+                Arguments.of(List.of(1, 2, 3, 4, 5, 6, 7))
+        );
+    }
 }
